Initial commit (functional, based on JSXC 4.1.1)
This commit contains a functional Openfire plugin that adds the JSXC 4.1.1 client.

The web client will be hosted on the same web server as that's used for Openfire's BOSH / http-bind endpoint, using the 'jsxc' context root. This makes the web client available on most servers on https://exmple.org:7483/jsxc/

Configuration options to be used by the web client are configurable in the Openfire admin console, which exposes them through an endpoint at /jsxc/options.json